Move update code to command
